% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLSSEM.R
\name{BLSSEM}
\alias{BLSSEM}
\title{This function performs BLSS.}
\usage{
BLSSEM(y, clin, X, s0, s1, error = 0.01, maxiter = 2000)
}
\arguments{
\item{y}{response variable}

\item{clin}{clinical factors}

\item{X}{genetic variables}

\item{s0}{spike scale}

\item{s1}{slab scale}

\item{error}{cutoff value}

\item{maxiter}{max iteration number}
}
\description{
This function performs BLSS.
}
