% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQLSSEM.R
\name{BQLSSEM}
\alias{BQLSSEM}
\title{This function performs BQLSS.}
\usage{
BQLSSEM(y, clin, X, quant, s0, s1, error = 0.01, maxiter = 2000)
}
\arguments{
\item{y}{response variable}

\item{clin}{clinical factors}

\item{X}{genetic variables}

\item{quant}{quantile}

\item{s0}{spike scale}

\item{s1}{slab scale}

\item{error}{cutoff value}

\item{maxiter}{max iteration number}
}
\description{
This function performs BQLSS.
}
