% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{cv}
\alias{cv}
\title{This function performs cross-validation.}
\usage{
cv(y, clin, X, quant, t0, t1, k, func, error, maxiter)
}
\arguments{
\item{y}{response variable}

\item{clin}{clinical factors}

\item{X}{genetic variables}

\item{quant}{quantile}

\item{t0}{spike scale range}

\item{t1}{slab scale range}

\item{k}{fold number}

\item{func}{choose function "BLSS" or "BQLSS"}

\item{error}{cutoff value}

\item{maxiter}{max iteration number}
}
\description{
This function performs cross-validation.
}
