% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ssem.R
\name{cv.ssem}
\alias{cv.ssem}
\title{k-folds cross-validation for SSEM}
\usage{
cv.ssem(y, clin, X, quant, t0, t1, k, func, error = 0.01, maxiter = 2000)
}
\arguments{
\item{y}{response variable}

\item{clin}{clinical factors}

\item{X}{genetic variables}

\item{quant}{quantile}

\item{t0}{spike scale range}

\item{t1}{slab scale range}

\item{k}{fold number}

\item{func}{choose function "BLSS" or "BQLSS"}

\item{error}{cutoff value}

\item{maxiter}{max iteration number}
}
\value{
a list with components:
\item{rs0}{the optimal spike scale under MAE.}
\item{rs1}{the optimal slab scale under MAE.}
\item{nrs0}{the optimal slab scale under MSE.}
\item{nrs1}{the optimal slab scale under MSE.}
\item{rCV}{cross-validation error matrix under MAE.}
\item{nrCV}{cross-validation error matrix under MSE.}
}
\description{
This function performs cross-validation.
}
\details{
When performing cross-validation for SSEM, function cv.SSEM returns two sets of optimal tuning parameters and their corresponding cross-validation error matrices. 
The spike scale parameter \eqn{rs0} and the slab scale parameter \eqn{rs1} are obtained based on the mean absolute error (MAE). 
The spike scale parameter \eqn{nrs0} and the slab scale parameter \eqn{nrs1} are obtained based on the mean squared error (MSE). 
Corresponding error matrices \eqn{rCV} and \eqn{nrCV} can also be obtained from the output.
}
