% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emBayes.R
\name{print.cv.emBayes}
\alias{print.cv.emBayes}
\title{print an cv.emBayes result}
\usage{
\method{print}{cv.emBayes}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{cv.emBayes result}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\value{
Print a list of output from a cv.emBayes object.
}
\description{
Print a summary of an 'cv.emBayes' result
}
\seealso{
\code{\link{cv.emBayes}}
}
