% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emBayes.R
\name{print.emBayes}
\alias{print.emBayes}
\title{print an emBayes result}
\usage{
\method{print}{emBayes}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{emBayes result}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\value{
Print a list of output from a emBayes object.
}
\description{
Print a summary of an 'emBayes' result
}
\seealso{
\code{\link{emBayes}}
}
