% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSEM.R
\name{SSEM}
\alias{SSEM}
\title{This function performs BQLSS.}
\usage{
SSEM(y, clin, X, quant, s0, s1, func, error = 0.01, maxiter = 2000)
}
\arguments{
\item{y}{response variable}

\item{clin}{clinical factors}

\item{X}{genetic variables}

\item{quant}{quantile}

\item{s0}{spike scale}

\item{s1}{slab scale}

\item{func}{choose function "BLSS" or "BQLSS"}

\item{error}{cutoff value}

\item{maxiter}{max iteration number}
}
\description{
This function performs BQLSS.
}
